Partial re-write of Storage.
Apparently Serializables within Serializables are not honored when
injecting into a SQL table so we must handle this ourselves.

The new format of persistence is Banks are a wrapper around Holder, World,
Account(s). This means the table now looks like the following:

-----------------------------------------------
| ID | Holder | World | AccountName | Balance |
-----------------------------------------------

When retrieving banks from the DB, we re-interpert them back into a
wrapper with accounts.
Signed-off-by: Chris Sanders <zidane@outlook.com>