Added ability to specify an interest payable.
This commit adds on the ability to designate an account within a bank as
the interest payable. When an interest clock occurs (once every 4 hrs
currently), the interest payable will div out payments. IF the interest
payable account is null (as in not set), it will default to the bank's
master balance.

Also API has been added to Bank to find if the bank has an interest
payable as well as gets the account.
Signed-off-by: Chris Sanders <zidane@outlook.com>